backend/requirements.txt
# Core Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0

# AI/ML Libraries
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Video Processing
av>=11.0.0
decord>=0.6.0
opencv-python>=4.8.0
yt-dlp>=2023.12.30

# Database & Storage
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0

# Monitoring & Logging
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
